Separate bootstrap integration test
